ChatBot/requirements.txt
numpy
streamlit
python-dotenv
google-generativeai